fix(beast-sim,beast-ecs): S11.2 review round-1 — relocate math, OnceLock statics
Address PR #259 review (4 MEDIUM + 5 LOW).

* MEDIUM (LOW #8 layering) — Move pure math (compute_engagement,
  compute_exposure, slot_adjacency, baseline arrays, gap/shield
  constants) from beast-ecs (L3 data foundation) to a new
  `beast_sim::formation` module (L4 domain logic). beast-ecs keeps
  only the data types (Formation, FormationSlot) and a data-only
  `Formation::from_slots` constructor. Per CRATE_LAYOUT.md, system
  logic belongs in L4.

* MEDIUM — `Formation::recompute` no longer exists on the data type;
  replaced with `beast_sim::formation::recompute(&mut Formation)`
  free function. The new `beast_sim::formation::build(slots)` wraps
  `from_slots` + `recompute` for the one-shot case.

* MEDIUM — `#[derive(Default)]` removed from Formation. A
  default-constructed Formation would have engagement/exposure at
  Q3232::ZERO (uncomputed state); removing the derive forces explicit
  construction via `from_slots` + `recompute` and surfaces the
  initialisation step in the type signature.

* MEDIUM — Replace `unwrap_or(0)` on adjacency lookups with
  `unwrap_or_else(|| panic!(...))`. Today this never fires (the table
  has all five keys), but silently treating a missing slot as having
  zero neighbours would mask any future SLOT_COUNT change that
  forgets to update `slot_adjacency()`.

* MEDIUM — Replace per-call BTreeMap allocation in `slot_adjacency()`
  + per-call `Q3232::from_num` array rebuild in `baseline_engagement`
  / `baseline_exposure` with `OnceLock`-backed accessor functions.
  `LazyLock` would be cleaner but the workspace MSRV is 1.75 and
  `LazyLock` only stabilised in 1.80; `OnceLock` is the existing
  pattern in beast-chronicler.

* LOW — Add adjacency rationale comment explaining why flank-L/R
  aren't directly adjacent and why vanguard isn't adjacent to center
  (rank-ordered topology per doc §3 — lateral coordination flows
  through center as intermediary).

* LOW — Document `terrain_modifier`'s expected `[-1, 1]` range.
  Out-of-range values still produce correct clamped output but waste
  intermediate work pushing the Q3232 sum towards saturation.

* LOW — Reframe `SLOT_NAMES` doc as "structural diagnostic
  identifiers" rather than "UI labels" — kept in beast-ecs for now;
  consumers in higher layers may localise.

* LOW — Replace accumulator `for _ in 0..n { x += step }` with
  `step * Q3232::from_num(n as i32)`. Q3232 multiplication is
  saturating like addition, and `n` is bounded to `0..=4` so there's
  no conversion ambiguity.

Test split: 5 storage / shape / SLOT_COUNT tests stay in
`beast-ecs::components::formation::tests`; 15 math + determinism
tests move to `beast-sim::formation::tests`. All 20 tests pass.
Determinism property (byte-identical Q32.32 output for same input)
preserved.
